backend/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.9
#psycopg2
python-dotenv==0.20.0
djangorestframework_simplejwt==5.1.0
cryptography==36.0.2
django-admin-list-filter-dropdown==1.0.3
gunicorn==20.1.0
sentry-sdk==1.5.8
django-otp==1.1.3
qrcode==7.3.1
docutils==0.18.1
requests==2.27.1
django-redis==5.2.0
django-filter==21.1
djongo==1.3.6
pymongo==3.12.1 # djongo 1.3.6 just compatible with pymongo
celery==5.2.7
django-celery-results==2.3.1
django-celery-beat==2.3.0
python-memcached==1.59
django-jalali-date==1.0.2
django-admin-rangefilter==0.10.0
django-better-admin-arrayfield==1.4.2
django-json-widget==1.1.1
fcm-django==2.0.0
pymemcache==4.0.0
model_bakery==1.17.0
django-cors-headers